Better separation of logical interfaces.
